Move CSV utils to gm4.utils
